# Control memory usage of downloads (including code-gen changes)